requirements.txt
Django==2.0.7
django-tastypie==0.14.1
ecdsa==0.13
python-dateutil==2.7.3
python-mimeparse==1.6.0
six==1.11.0
